// src_fatcAng/app/shared/data.dto.ts
export interface ICollectionSummary {
  id: number;
  caseId: number;
  collectorId: number;
  collectorSummary: { [itemName: string]: string; };
  retrievedAt: Date;
}

export interface ICollectorConfiguration {
  id: number;
  caseID: number;
  collectorTypeName: string;
  collectorAlias: string;
  hasAuthorization: boolean;
  authInfo: IAuthorization;
  desiredChunkSize: number;
  indexNumber: number;
  authRequestID: string;
  summary: ICollectionSummary;
}

export interface IAuthorization {
  username: string;
  password: string;
  authCode: string;
  authRequestID: string;
}

export interface ICaseConfig {
  case: ICaseDTO;
  configurations: ICollectorConfiguration[];
}

export interface ICaseCollectionHistory {
  id: number;
  runID: string;
  collectorConfig: ICollectorConfiguration;
  ownerID: number;
  startedAt: number; // epochs
  finishedAt: number;
  status: number;
  errorMessage: string;
  sizeBytes: number;
  filename: string;
}

export const WS_PROGRESS = "progress";
export const WS_LINK = "link";
export const WS_LOG = "log";
export const WS_COLLECTION_DONE = "collection_done";

// websocket/msg.go
export interface IWebsocketMessage {
  type:  "progress" | "link" | "log" | "collection_done";
  version: number;
  caseID: number;
  collectorConfigID: number;
}